Fixed Sim-environment-current-volume-malfunction-scaling bug.
Sim environment permanent malfunctions are implemented by scaling
down the current volume and thus its atmospheric store.  For a
severe permanent malfunction it is set to 0 while a medium one
reduces it by 50%.  The original intent for low permanent
malfunctions appears to have been reduction by 25%, but as
implemented it is reduced to (not by) 25%, resulting in a low
malfunction being more severe than a medium one.

* Corrected current volume scaling for permanent low malfunctions.