Fixed Sim-Environment-malfunction-gas-store-adjustment bug.
When the current volume is reduced due to permanent malfunction, the
individual gas stores are adjusted accordingly.  While doing so,
setCurrentLevel(), which expects to receive moles, is given liters.
This results in significantly more gas than is correct -- 24.3 times
as much based on moleOfHumidAirPerLiter = 0.04115f at 23C and 101kPa
used to initialize the gas stores.

* Following permanent malfunction current volume reduction, adjust
  current gas store levels using moles.